back-end/pyproject.toml
[tool.poetry]
name = "test-fastapi"
version = "0.1.0"
description = "FastAPI-Test"
authors = ["MIT"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.78.0"
pytest = "^7.1.2"
uvicorn = "^0.17.6"
pandas = "^1.4.2"
pydantic = "^1.9.1"
PyYAML = "^6.0"
pymongo = {extras = ["srv"], version = "^4.1.1"}
requests = "^2.27.1"
joblib = "^1.1.0"
numpy = "^1.22.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
